requirements.txt
Flask==2.3.3
pymongo==4.5.0
flask-login==0.6.3
flask-wtf==1.1.1
werkzeug==2.3.7
python-dotenv==1.0.0
flask-mail==0.9.1
bcrypt==4.0.1
schedule==1.2.0
